+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # master
 
+* Allow the transitions / known states for an event to be reset
 * Add fire_#{name}_event instance method for firing an arbitrary event on a state machine
 * Improve InvalidTransition exception messages to include the failure reason(s) in ORM integrations
 * Don't allow around_transitions to attempt to be called in multiple execution contexts when run in jruby